finish exercise 4.21

+++ part4/blog/controllers/blogs.js
@@ -17,9 +17,12 @@ blogsRouter.get('/:id', async(request, response) => {
 })
   
 blogsRouter.post('/', async (request, response) => {
+    if (!request.token) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    
+
     if (!decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
@@ -34,9 +37,28 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   const id = request.params.id
-  await Blog.findByIdAndRemove(id)
-  response.status(204).end()
+
+  const blog = await Blog.findById(id)
+
+  if(blog.user.toString() === decodedToken.id.toString()){
+    await Blog.findByIdAndRemove(id)
+    response.status(204).end()
+  }else{
+    response.status(401).json({ error: 'Insufficient permissions' })
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
